rename and restructure commit coverage
