+++ src/pages/CommitDetailPage/CommitCoverage/index.js
@@ -1,0 +1,1 @@
+export { default } from './CommitCoverage'
