--- src/pages/CommitDetailPage/CommitDetailPageContent/index.js
@@ -1,1 +1,0 @@
-export { default } from './CommitDetailPageContent'
